% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{argonpkg-package}
\alias{argonpkg}
\alias{argonpkg-package}
\title{argonpkg: An Example of a golem app wrapping an argon dashboard template..}
\description{
An Example of a golem app wrapping an argon dashboard template.
}
\author{
\strong{Maintainer}: Ernest Omane-Kodie \email{eokodie@gmail.com}

}
